requirements.txt
pyHS100==0.3.5
paho-mqtt==1.5.0
